Fixed simulation data generator - I discovered that the API does not handle multiple transitions on the same sample index properly. Updated CMakeLists.txt for Windows and Linux, however it has only been tested on Windows.

+++ source/AsyncRgbLedAnalyzerSettings.cpp
@@ -40,7 +40,7 @@ AsyncRgbLedAnalyzerSettings::~AsyncRgbLedAnalyzerSettings()
 bool AsyncRgbLedAnalyzerSettings::SetSettingsFromInterfaces()
 {
 	mInputChannel = mInputChannelInterface->GetChannel();
-	mLEDController = static_cast<Controller>(mControllerInterface->GetNumber());
+	mLEDController = static_cast<Controller>( static_cast<int>(mControllerInterface->GetNumber()));
 
 	ClearChannels();
 	AddChannel( mInputChannel, DEFAULT_CHANNEL_NAME, true );
